video-encoder/requirements.txt
git+https://github.com/openai/CLIP.git
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.11
ftfy==6.1.1
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-bigquery==2.32.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.43.0
idna==3.3
numpy==1.21.5
packaging==21.3
Pillow==9.0.1
proto-plus==1.20.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
regex==2022.1.18
requests==2.27.1
rsa==4.8
six==1.16.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.62.3
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
